+++ doc/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 
+- Hightower servers is now dead, its only LSM now (updated readme).
 ### Changed
 
 ## [1.0.0] 9 September 2026