+++ doc/CHANGELOG.md
@@ -8,6 +8,15 @@
 
 ### Changed
 
+## [1.0.1] 18 September 2026
+
+### Added
+
+### Fixed
+
+- Hightower servers is now dead, its only LSM now (updated readme).
+### Changed
+
 ## [1.0.0] 9 September 2026
 
 ### Added